feat: use integer math

+++ dtmf/src/wave.h
@@ -5,9 +5,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
-int wave_generate(const char *path, float *buffer, size_t len,
+int wave_generate(const char *path, int32_t *buffer, size_t len,
 		  uint32_t channels, uint32_t sample_rate);
 
-float *wave_read(const char *path, size_t *len, double *sample_rate);
+int32_t *wave_read(const char *path, size_t *len, double *sample_rate);
 
 #endif
